Refactored shell integration test for easily actors setup to common/mod.rs
